oa_diagnosis/requirements.txt
pyautogen>=0.2.0
numpy
pandas
pydicom
openai
chainlit
